♻️ Refactor generate select template to isolate templated code to the minimum
